requirements.txt
tk
python-docx
docx2txt